+++ src/lexical/error.cc
@@ -1,3 +1,4 @@
+#include <stdio.h>
 #include "index.h"
 #include "../utils/index.h"